feat: 当用户还未配置行列头和数值时，不渲染表格框架
