+++ packages/s2-core/src/facet/header/util.ts
@@ -1,5 +1,4 @@
-import { isEmpty, isNil } from 'lodash';
-import type { Fields } from '../../common/interface';
+import { isNil } from 'lodash';
 
 export const getCellPadding = () => {
   const padding = [12, 4, 12, 4];
@@ -13,18 +12,4 @@ export const getCellPadding = () => {
     top,
     bottom,
   };
-};
-
-/**
- * fields 的 rows、columns、values、customTreeItems 值都为空时，返回 true
- * @param {Fields} fields
- * @return {boolean}
- */
-export const areAllFieldsEmpty = (fields: Fields) => {
-  return (
-    isEmpty(fields.rows) &&
-    isEmpty(fields.columns) &&
-    isEmpty(fields.values) &&
-    isEmpty(fields.customTreeItems)
-  );
 };
